MultiIndexing

Documentation for MultiIndexing.jl.

MultiIndexing.FixedMultiIndexSet — Type

A FixedMultiIndexSet is a sparse representation of a multi-index set, where we only keep track of nonzero values.

starts: Vector{Int} - the starting index of each multi-index nzindices: Vector{Int} - the nonzero indices of each multi-index nzvalues: Vector{Int} - the nonzero values of each multi-index

nzindices and nzvalues will have the same length M, and if there are N multi-indices, then length(starts) == N + 1. By construction, starts[1] == 1 and starts[N + 1] == M + 1.

MultiIndexing.MultiIndexSet — Type

Represents a set of multi-indices vector of {d} dimensional multi-indices with length N

MultiIndexing.MultiIndexSet — Method

MultiIndexSet(indices, limit, calc_reduced_margin)

Create a multi-index set from a matrix of multi-indices

Arguments:

indices: dim x N matrix of multi-indices

limit: function that takes a multi-index and a limit and returns true if the index is admissible

calc_reduced_margin: if true, calculate the reduced margin of the set, otherwise leave it empty

MultiIndexing.CreateTensorOrder — Method

CreateTensorOrder(d, p, limit)

Create a multi-index set with tensor order p

MultiIndexing.CreateTotalOrder — Function

CreateTotalOrder(d, p, limit)

Create a multi-index set with total order p

MultiIndexing.SmolyakQuadrature — Method

SmolyakQuadrature(mset, rules)

Create a Smolyak quadrature rule from a multi-index set and a set of rules

Arguments

mset: MultiIndexSet

rules: Vector of rules for each dimension. rules[j](n::Int) should return a quadrature rule (pts,wts) for dimension j exact up to order n

MultiIndexing.allBackwardAncestors — Method

allBackwardAncestors(mis, j)

Get all indices of multi-indices in mset limited by the tensor product box of the mset at idx

Arguments:

mis: MultiIndexSet to get the backward ancestors from

j: index of the multi-index to get the backward ancestors of

julia> d, p = 2, 5;

julia> mis = CreateTotalOrder(d, p);

julia> println(visualize_2d(mis))
X
X X
X X X
X X X X
X X X X X
X X X X X X

julia> midx = [2,2]; idx = findfirst(isequal(midx), vec(mis.indices));

julia> ancestors_idx = allBackwardAncestors(mis, idx);

julia> ancestors = MultiIndexSet(mis[ancestors_idx]);

julia> println(visualize_2d(ancestors))
X X
X X X
X X X

MultiIndexing.checkIndexAdmissible — Method

checkIndexAdmissible(mis, idx, check_indices)

Check if an index is admissible to add to a reduced margin

MultiIndexing.findReducedFrontier — Method

findReducedFrontier(mis)

Find the subset of multi-indices that are not the backward neighbors of any other multi-index in the set.

Examples

julia> d, p = 2, 3;

julia> mis = CreateTotalOrder(d, p);

julia> frontier = findReducedFrontier(mis);

julia> length(frontier)
4

julia> expected_indices = [[0,3], [1,2], [2,1], [3,0]];

julia> all(e in vec(mis) for e in expected_indices)
true

MultiIndexing.smolyakIndexing — Method

smolyakIndexing(mset) -> Vector{Tuple{Int,Int}}

Gets the smolyak indexing for a multi-index set. Returns a vector of tuples with each first index as the index of the multi-index representing a tensor product rule and each second index representing its count in the smolyak construction.

Examples

julia> d = 2;

julia> mis = MultiIndexSet([(0:7)'; zeros(Int,1,8)]);

julia> quad_rules = smolyakIndexing(mis); # Creates quad rule exact on x^7

julia> quad_rules[1], length(quad_rules) # Counts the highest index once
((8, 1), 1)

julia> mis = CreateTotalOrder(d, 10);

julia> print(visualize_smolyak_2d(mis, false))
X
X X
o X X
o o X X
o o o X X
o o o o X X
o o o o o X X
o o o o o o X X
o o o o o o o X X
o o o o o o o o X X
o o o o o o o o o X X

MultiIndexing.visualize_2d — Function

visualize_2d(mset_mat, markers)

Visualize a 2D multi-index set as a string with markers for each index.

Arguments:

mset: MultiIndexSet{2} or 2 x N matrix of multi-indices

markers: a character or array of characters (length N) to use as markers for each index

Examples

julia> mis = CreateTotalOrder(2, 4);

julia> println(visualize_2d(mis))
X
X X
X X X
X X X X
X X X X X

MultiIndexing.visualize_smolyak_2d — Function

visualize_smolyak_2d(mset)

Visualize a two-dimensional mset's smolyak decomposition with colors!

Examples

julia> mis = MultiIndexSet([0 1 4 3 2 1 0 0 0 0; 0 1 0 0 0 0 1 2 3 4]);

julia> println(visualize_smolyak_2d(mis, false))
X
o
o
X X
o X o o X